fix(api): setup for e2e tests
