+++ packages/shared/lib/db/seeders/account.seeder.ts
@@ -1,6 +1,5 @@
 import { v4 as uuid } from 'uuid';
 import type { Account } from '../../models/index.js';
-
 import accountService from '../../services/account.service.js';
 
 export async function createAccount(): Promise<Account> {
