+++ packages/shared/lib/db/seeders/index.ts
@@ -1,5 +1,8 @@
-export * from './environment.seeder.js';
+export * from './account.seeder.js';
 export * from './config.seeder.js';
 export * from './connection.seeder.js';
-export * from './sync.seeder.js';
+export * from './environment.seeder.js';
+export * from './global.seeder.js';
 export * from './sync-job.seeder.js';
+export * from './sync.seeder.js';
+export * from './user.seeder.js';
