+++ packages/types/lib/api.ts
@@ -1,15 +1,26 @@
-export interface ApiError<TCode extends string> {
+export interface ApiError<TCode extends string, TErrors = undefined> {
     error: {
         code: TCode;
         message?: string | undefined;
+        errors?: TErrors;
     };
 }
+export interface ValidationError {
+    code: string;
+    message: string;
+    path: (string | number)[];
+}
 
+export type ResDefaultErrors = ApiError<'not_found'> | ApiError<'invalid_query_params', ValidationError[]> | ApiError<'invalid_body', ValidationError[]>;
+
+export type EndpointMethod = 'GET' | 'POST' | 'PATCH' | 'PUT' | 'DELETE';
 /**
  * API Request/Response type
  */
 export interface Endpoint<
     T extends {
+        Method: EndpointMethod;
+        Path: string;
         Params?: Record<string, any>;
         Body?: Record<string, any>;
         Querystring?: Record<string, any>;
@@ -21,6 +32,8 @@ export interface Endpoint<
 > {
     // ------------
     // ------------ Request
+    Method: T['Method'];
+    Path: T['Path'];
     /**
      * URL params
      */
@@ -56,5 +69,5 @@ export interface Endpoint<
     /**
      * Response body (success + error)
      */
-    Reply: T['Error'] extends ApiError<any> ? T['Error'] | T['Success'] : T['Success'];
+    Reply: ResDefaultErrors | (T['Error'] extends ApiError<any> ? T['Error'] | T['Success'] : T['Success']);
 }
